data_collection/pyproject.toml
[project]
name = "data_collection"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Aryan Iyappan", email = "aryan@velocius.ai" }]
dependencies = [
    "dotenv>=0.0.5",
    "langchain>=0.2.12",
    "langchain-openai>=0.1.20",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "azure-ai-documentintelligence>=1.0.0b3",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "typer>=0.12.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
process-lease-agreements = "python -m scripts.process_lease_agreements"
